+++ Engine_SOURCE/yaInput.cpp
@@ -77,8 +77,8 @@ namespace ya
 			POINT mousePos = {};
 			GetCursorPos(&mousePos);
 			ScreenToClient(application.GetHwnd(), &mousePos);
-			mMousePosition.x = mousePos.x;
-			mMousePosition.y = mousePos.y;
+			mMousePosition.x = (float)mousePos.x;
+			mMousePosition.y = (float)mousePos.y;
 		}
 		else
 		{
